Game-Seed association, aka better multiworld UX (tm)

+++ src/main/kotlin/wotw/server/database/model/Multiverses.kt
@@ -19,13 +19,13 @@ import kotlin.math.ceil
 import kotlin.to
 
 object Multiverses : LongIdTable("multiverse") {
-    override val primaryKey = PrimaryKey(id)
     val seed = reference("seed", Seeds).nullable()
     val board = jsonb("board", BingoCard.serializer()).nullable()
 }
 
 class Multiverse(id: EntityID<Long>) : LongEntity(id) {
     var board by Multiverses.board
+    var seed by Seed optionalReferencedOn Multiverses.seed
     val universes by Universe referrersOn Universes.multiverseId
     val worlds: Collection<World>
         get() = universes.flatMap { it.worlds }
@@ -169,11 +169,12 @@ class Multiverse(id: EntityID<Long>) : LongEntity(id) {
                 affectedMultiverseIds.add(it.universe.multiverse.id.value)
                 it.members = SizedCollection(it.members.minus(player))
 
-                if (it.members.empty()) {
+                //Do not delete empty worlds in universes which are attached to a seed
+                if (it.members.empty() && it.universe.multiverse.seed == null) {
                     it.delete()
                 }
 
-                if (it.universe.worlds.empty()) {
+                if (it.universe.worlds.all { it.members.empty() } && newWorld?.universe != it.universe) {
                     it.universe.delete()
                 }
             }
